+++ Configuration/TCA/Overrides/sys_template.php
@@ -7,7 +7,7 @@
  * LICENSE.txt file that was distributed with this source code.
  */
 
-defined('TYPO3_MODE') || die();
+defined('TYPO3') || die();
 
 TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addStaticFile(
     Brotkrueml\FormCountrySelect\Extension::KEY,